# search: Minor comby translation perf improvement